refactor: fetchApi로 추상화된 함수들을 useCart, useProduct에 적용
