+++ src/hooks/useCart.ts
@@ -2,14 +2,18 @@ import { useRecoilCallback, useRecoilValue } from 'recoil';
 import { cartState } from '../store/CartState';
 import { ChangeEvent, MouseEvent, useState } from 'react';
 import { removeProductItemFromCartSelector, totalPriceSelector } from '../store/CartSelector';
+import { CART_BASE_URL } from '../constants/url';
+import { useFetchData } from './useFetchData';
 
-export const useHandleCartList = () => {
+export const useCart = () => {
   const cart = useRecoilValue(cartState);
   const initialCheckedItems = cart.map((item) => item.id);
   const [checkedItems, setCheckedItems] = useState<number[]>(initialCheckedItems);
   const removeProductItemFromCart = useRecoilCallback(({ set }) => (id: number) => {
     set(removeProductItemFromCartSelector(id), []);
   });
+
+  const { api } = useFetchData();
 
   const isChecked = (id: number) => {
     return checkedItems.includes(id);
@@ -34,7 +38,10 @@ export const useHandleCartList = () => {
   };
 
   const handleRemoveCheckedItem = () => {
-    checkedItems.forEach((id) => removeProductItemFromCart(id));
+    checkedItems.forEach((id) => {
+      api.delete(`${CART_BASE_URL}/${id}`, { id });
+      removeProductItemFromCart(id);
+    });
     setCheckedItems([]);
   };
